[fea-rs] Rename gsub/gpos module to gsub/gpos{_builders}
The shorter names were okay when fea-rs was in its repo, but it was
getting annoying to use things like fzf to try and find them here.
